+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add [d3 legend](https://d3-legend.susielu.com/) library for usage in examples.
 - Add subtitles to axes.
 - Add cartesian chart component to handle flipping of axes.
+- Add legend component.
+- Add square symbol legend data function.
 
 ### Changed